reorganize for packaging
